Move dataclasses into `__init__.py` in order to avoid circular dependencies
Colocating dataclasses are colocated in the same file as most of your
code, this makes splitting code into different files difficult, as one
cannot introduce new code which makes use of the datamodels, and is used
by the colocated code. This introduces circular dependencies.
